minmax inject to math
