+++ src/template/Math.hpp
@@ -109,3 +109,88 @@ namespace math {
     }
 
 } // namespace math
+
+/**
+ * @brief 2項の最小値求める
+ * @tparam T
+ */
+template <typename T>
+struct minFunc {
+    /**
+     * @brief 本体
+     * @param l
+     * @param r
+     * @return T
+     */
+    static T exec(const T l, const T r) { return l < r ? l : r; }
+};
+
+/**
+ * @brief 2項の最大値求める
+ *
+ * @tparam T
+ */
+template <typename T>
+struct maxFunc {
+    /**
+     * @brief 本体
+     *
+     * @param l
+     * @param r
+     * @return T
+     */
+    static T exec(const T l, const T r) { return l > r ? l : r; }
+};
+
+/**
+ * @brief 複数項の最小値
+ * @see chain
+ * @tparam T
+ * @tparam Ts
+ * @param head
+ * @param tail
+ * @return T
+ */
+template <typename T, typename... Ts>
+inline T minOf(T head, Ts... tail) {
+    return chain<minFunc<T>>(head, tail...);
+}
+
+/**
+ * @brief 複数項の最大値
+ * @see chain
+ * @tparam T
+ * @tparam Ts
+ * @param head
+ * @param tail
+ * @return T
+ */
+template <typename T, typename... Ts>
+inline T maxOf(T head, Ts... tail) {
+    return chain<maxFunc<T>>(head, tail...);
+}
+
+/**
+ * @brief change min
+ * @tparam T 型
+ * @param target 変更する値
+ * @param candidates
+ * @return 更新があればtrue
+ */
+template <typename T, typename... Ts>
+inline bool chmin(T &target, Ts &&... candidates) {
+    return changeTarget<minFunc<T>>(target, candidates...);
+}
+
+/**
+ * @brief chminのmax版
+ * @see chmin
+ * @tparam T 型
+ * @param target 変更する値
+ * @param candidates
+ * @return 更新があればtrue
+ */
+template <typename T, typename... Ts>
+inline bool chmax(T &target, Ts &&... candidates) {
+    return changeTarget<maxFunc<T>>(target, candidates...);
+}
