+++ test/template/CMakeLists.txt
@@ -1,5 +1,3 @@
-add_executable(MinMaxOperation.test MinMaxOperation.test.cpp)
-add_test(NAME MinMaxOperation COMMAND MinMaxOperation.test)
 add_executable(Random.test Random.test.cpp)
 add_test(NAME Random COMMAND Random.test)
 add_executable(Macro.test Macro.test.cpp)
